samples changes, mvp project structure
