+++ archx/src/main/java/co/en/archx/archx/medium/Event.java
@@ -1,4 +1,4 @@
-package co.en.archx.archx.transferobjects;
+package co.en.archx.archx.medium;
 
 public interface Event<A extends Action> {
