+++ archx/src/main/java/co/en/archx/archx/medium/Result.java
@@ -1,0 +1,3 @@
+package co.en.archx.archx.medium;
+
+public interface Result { }
