--- archx/src/main/java/co/en/archx/archx/transferobjects/Result.java
@@ -1,3 +1,0 @@
-package co.en.archx.archx.transferobjects;
-
-public interface Result { }
