--- sample/src/main/java/co/en/archx/sample/screens/main/MainView.kt
@@ -1,11 +1,0 @@
-package co.en.archx.sample.screens.main
-
-import co.en.archx.archx.IView
-import co.en.archx.sample.screens.main.transferobjects.MainEvent
-import co.en.archx.sample.screens.main.transferobjects.MainState
-
-interface MainView : IView<MainEvent, MainState> {
-
-
-
-}
