--- sample/src/main/java/co/en/archx/sample/screens/main/transferobjects/MainAction.kt
@@ -1,7 +1,0 @@
-package co.en.archx.sample.screens.main.transferobjects
-
-import co.en.archx.archx.transferobjects.Action
-
-sealed class MainAction : Action {
-
-}
