--- sample/src/main/java/co/en/archx/sample/screens/main/transferobjects/MainEvent.kt
@@ -1,7 +1,0 @@
-package co.en.archx.sample.screens.main.transferobjects
-
-import co.en.archx.archx.transferobjects.Event
-
-sealed class MainEvent : Event<MainAction> {
-
-}
