--- sample/src/main/java/co/en/archx/sample/screens/main/transferobjects/MainResult.kt
@@ -1,7 +1,0 @@
-package co.en.archx.sample.screens.main.transferobjects
-
-import co.en.archx.archx.transferobjects.Result
-
-sealed class MainResult : Result {
-
-}
