--- sample/src/main/java/co/en/archx/sample/screens/main/transferobjects/MainState.kt
@@ -1,12 +1,0 @@
-package co.en.archx.sample.screens.main.transferobjects
-
-import co.en.archx.archx.transferobjects.State
-
-class MainState() : State {
-
-    companion object {
-
-        fun init() = MainState()
-    }
-
-}
