+++ sample/src/main/java/co/en/archx/sample/ui/activity/main/MainView.kt
@@ -1,0 +1,11 @@
+package co.en.archx.sample.ui.activity.main
+
+import co.en.archx.archx.IView
+import co.en.archx.sample.ui.activity.main.transferobjects.MainEvent
+import co.en.archx.sample.ui.activity.main.transferobjects.MainState
+
+interface MainView : IView<MainEvent, MainState> {
+
+
+
+}
