+++ sample/src/main/java/co/en/archx/sample/ui/activity/main/transferobjects/MainAction.kt
@@ -1,0 +1,7 @@
+package co.en.archx.sample.ui.activity.main.transferobjects
+
+import co.en.archx.archx.medium.Action
+
+sealed class MainAction : Action {
+
+}
