+++ sample/src/main/java/co/en/archx/sample/ui/activity/main/transferobjects/MainEvent.kt
@@ -1,0 +1,7 @@
+package co.en.archx.sample.ui.activity.main.transferobjects
+
+import co.en.archx.archx.medium.Event
+
+sealed class MainEvent : Event<MainAction> {
+
+}
