+++ sample/src/main/java/co/en/archx/sample/ui/activity/main/transferobjects/MainResult.kt
@@ -1,0 +1,7 @@
+package co.en.archx.sample.ui.activity.main.transferobjects
+
+import co.en.archx.archx.medium.Result
+
+sealed class MainResult : Result {
+
+}
