+++ sample/src/main/java/co/en/archx/sample/ui/activity/main/transferobjects/MainState.kt
@@ -1,0 +1,12 @@
+package co.en.archx.sample.ui.activity.main.transferobjects
+
+import co.en.archx.archx.medium.State
+
+class MainState() : State {
+
+    companion object {
+
+        fun init() = MainState()
+    }
+
+}
